+++ src/django_feature_policy.py
@@ -63,7 +63,7 @@ FEATURE_NAMES = {
 }
 
 
-class FeaturePolicyMiddleware:
+class PermissionsPolicyMiddleware:
     def __init__(self, get_response):
         self.get_response = get_response
         self.header_value  # Access at setup so ImproperlyConfigured can be raised
@@ -73,12 +73,13 @@ class FeaturePolicyMiddleware:
         response = self.get_response(request)
         value = self.header_value
         if value:
-            response["Feature-Policy"] = value
+            response["Permissions-Policy"] = value
+            response["Feature-Policy"] = self.old_header_value
         return response
 
     @cached_property
     def header_value(self):
-        setting = getattr(settings, "FEATURE_POLICY", {})
+        setting = self.get_setting()
         pieces = []
         for feature, values in sorted(setting.items()):
             if feature not in FEATURE_NAMES:
@@ -86,20 +87,53 @@ class FeaturePolicyMiddleware:
             if isinstance(values, str):
                 values = (values,)
 
-            item = [feature]
+            item = []
             for value in values:
                 if value == "none":
-                    item.append("'none'")
-                elif value == "self":
-                    item.append("'self'")
-                else:
+                    # 'none' was previously supported as a special token for
+                    # Feature-Policy, now can be represented by the empty list.
+                    pass
+                elif value in ("self", "*"):
                     item.append(value)
+                else:
+                    item.append('"{}"'.format(value))
+            pieces.append(feature + "=(" + " ".join(item) + ")")
+        return ", ".join(pieces)
+
+    @cached_property
+    def old_header_value(self):
+        setting = self.get_setting()
+        pieces = []
+        for feature, values in sorted(setting.items()):
+            if isinstance(values, str):
+                values = (values,)
+
+            item = [feature]
+            if not values:
+                item.append("'none'")
+            else:
+                for value in values:
+                    if value == "none":
+                        item.append("'none'")
+                    elif value == "self":
+                        item.append("'self'")
+                    else:
+                        item.append(value)
             pieces.append(" ".join(item))
         return "; ".join(pieces)
 
+    def get_setting(self):
+        setting = getattr(settings, "PERMISSIONS_POLICY", None)
+        if not setting:
+            setting = getattr(settings, "FEATURE_POLICY", {})
+        return setting
+
     def clear_header_value(self, setting, **kwargs):
-        if setting == "FEATURE_POLICY":
+        if setting in ("PERMISSIONS_POLICY", "FEATURE_POLICY"):
             try:
                 del self.header_value
             except AttributeError:
                 pass
+
+
+FeaturePolicyMiddleware = PermissionsPolicyMiddleware